[mcrecover] CardPrivate::open(): Cast to quint64 before multiplying out the file size.
This ensures the 32-bit multiplication result doesn't overflow before
the result is stored in the 64-bit filesize variable.

This fixes Coverity#1391605.

+++ src/mcrecover/card/Card.cpp
@@ -140,7 +140,7 @@ int CardPrivate::open(const QString &filename, QIODevice::OpenModeFlag mode)
 		// Too many blocks.
 		// Only read up to maxBlocks.
 		this->errors |= Card::MCE_SZ_TOO_BIG;
-		this->filesize = (maxBlocks * blockSize);
+		this->filesize = (static_cast<quint64>(maxBlocks) * blockSize);
 	}
 
 	if (!isPow2(this->filesize)) {
